make pr changes from comments
