+++ src/signature_mahalanobis_knn/utils.py
@@ -4,6 +4,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 from sklearn.metrics import roc_auc_score, roc_curve
 
+from signature_mahalanobis_knn.sig_mahal_knn import SignatureMahalanobisKNN
+
 
 def plot_roc_curve(
     y_true: np.ndarray,
@@ -48,3 +50,114 @@ def plot_roc_curve(
     plt.show()
 
     return roc_auc
+
+
+def compute_auc_given_dists(
+    distances_in: np.ndarray,
+    distances_out: np.ndarray,
+    plot: bool = False,
+    title: str = "",
+) -> float:
+    """
+    Compute ROC AUC given the distances of inliers and outliers.
+
+    Parameters
+    ----------
+    distances_in : np.ndarray
+        KNN distances for the inlier data points.
+    distances_out : np.ndarray
+        KNN distances for the outlier data points.
+    plot : bool, optional
+        Whether to plot the ROC curve, by default False
+    title : str, optional
+        Title for the ROC curve plot, by default "".
+        Only used when plot is True.
+
+    Returns
+    -------
+    float
+        ROC AUC score.
+    """
+    # replace infinity with twice of the maximum value, hacky, may need more thoughts
+    distances_in[distances_in == np.inf] = np.nan
+    distances_out[distances_out == np.inf] = np.nan
+    max_val = max(np.nanmax(distances_in), np.nanmax(distances_out))
+    distances_in = np.nan_to_num(distances_in, max_val * 2)
+    distances_out = np.nan_to_num(distances_out, max_val * 2)
+
+    y_true = [0] * len(distances_in) + [1] * len(distances_out)
+    y_score = np.concatenate([distances_in, distances_out])
+    roc_auc = roc_auc_score(
+        y_true=y_true,
+        y_score=y_score,
+    )
+
+    if plot:
+        plot_roc_curve(
+            y_true=y_true,
+            y_score=y_score,
+            roc_auc=roc_auc,
+            title=title,
+        )
+
+    return roc_auc
+
+
+def compute_auc(
+    signature_mahalanobis_knn: SignatureMahalanobisKNN,
+    test_in: np.ndarray,
+    test_out: np.ndarray,
+    is_signature: bool,
+    plot: bool = False,
+    title: str = "",
+) -> float:
+    """
+    Compute ROC AUC given the data points of inliers and outliers.
+
+    Parameters
+    ----------
+    test_in : np.ndarray
+        Data points from the inlier class.
+    test_out : np.ndarray
+        Data points from the outlier class.
+    is_signature : bool
+        Whether the data provided are signatures or not.
+        If True, then test_in and test_out are signatures.
+        If False, then test_in and test_out are data points
+        and the signatures will be computed.
+    plot : bool, optional
+        Whether to plot the ROC curve, by default False
+    title : str, optional
+        Title for the ROC curve plot, by default "".
+        Only used when plot is True.
+
+    Returns
+    -------
+    float
+        ROC AUC score.
+    """
+    # compute KNN distances for the signatures of the data points
+    # for both inliers and outliers of distribution data
+    if is_signature:
+        distances_in = signature_mahalanobis_knn.conformance(signatures=test_in)
+        distances_out = signature_mahalanobis_knn.conformance(signatures=test_out)
+    else:
+        distances_in = signature_mahalanobis_knn.conformance(X=test_in)
+        distances_out = signature_mahalanobis_knn.conformance(X=test_out)
+
+    # convert to the default data type of the arrays in
+    # the mahalanobis distance object
+    distances_in = distances_in.astype(
+        signature_mahalanobis_knn.mahal_distance.default_dtype
+    )
+    distances_out = distances_out.astype(
+        signature_mahalanobis_knn.mahal_distance.default_dtype
+    )
+
+    # compute AUC for the inliers and outliers
+    return compute_auc_given_dists(
+        distances_in=distances_in,
+        distances_out=distances_out,
+        plot=plot,
+        title=title,
+    )
